+++ src/Listener/Hook/RegisterFixContentParentRelationsFixerListener.php
@@ -6,8 +6,8 @@ namespace ContaoBootstrap\Grid\Listener\Hook;
 
 use ContaoBootstrap\Grid\Listener\Dca\ParentFixContentParentRelationsListener;
 use Netzmacht\Contao\Toolkit\Assertion\AssertionFailed;
-use Netzmacht\Contao\Toolkit\Dca\Definition;
 use Netzmacht\Contao\Toolkit\Dca\DcaManager;
+use Netzmacht\Contao\Toolkit\Dca\Definition;
 
 use function in_array;
 use function is_array;