+++ src/Resources/contao/dca/tl_content.php
@@ -13,6 +13,11 @@ $GLOBALS['TL_DCA']['tl_content']['config']['onload_callback'][] = [
     'initializeDca',
 ];
 
+$GLOBALS['TL_DCA']['tl_content']['config']['onload_callback'][] = [
+    'contao_bootstrap.grid.listeners.dca.content',
+    'updatePaletteOnNestedParent',
+];
+
 $GLOBALS['TL_DCA']['tl_content']['config']['oncopy_callback'][] = [
     ContentFixParentRelationListener::class,
     'onCopy',
@@ -26,6 +31,20 @@ $GLOBALS['TL_DCA']['tl_content']['config']['onsubmit_callback'][] = [
 /*
  * Palettes
  */
+
+$GLOBALS['TL_DCA']['tl_content']['metapalettes']['bs_grid_wrapper'] = [
+    'type'           => [
+        'title',
+        'headline',
+        'type',
+        'bs_grid',
+        'bs_grid_name',
+    ],
+    'template'       => [':hide', 'customTpl'],
+    'protected'      => [':hide', 'protected'],
+    'expert'         => [':hide', 'guests', 'cssID'],
+    'invisible'      => ['invisible', 'start', 'stop'],
+];
 
 $GLOBALS['TL_DCA']['tl_content']['metapalettes']['bs_gridStart'] = [
     'type'           => [